+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - Display the GPU unique ID in the GPU info tab. Available on Vega and later GPU models.
 - Make custom profile icons optional in the profile dialog. The user can now select a custom icon or simply use the default one in their profiles (#127).
 - Manual profiles. The user can now create profiles to partially or completely override  the settings applied by the global and per-application profiles. They can be toggled from the system tray or through the main GUI (#49).
+- System tray menu entry to hide and show the main window (#280).
 - Add German translation. Thanks to Rüdiger Arp, polyphase and SE. (#104).